Move sim components into separate folder. Remove old sim footer. Move sim outcome into bottom video.
