+++ src/components/new-simulation/button-toggle.scss
@@ -1,4 +1,4 @@
-@use "./vars.scss";
+@use "../vars.scss";
 
 .button-row {
   display: flex;
