+++ src/components/new-simulation/simulation-settings.scss
@@ -1,4 +1,4 @@
-@use "./vars.scss";
+@use "../vars.scss";
 
 .simulation-settings {
   border: vars.$sim-border;
@@ -44,21 +44,21 @@
       margin: 8.5px 0 5px -10px;
 
       .play-button {
-        background-image: url("../assets/new-sim/buttons/play-button-on.svg");
+        background-image: url("../../assets/new-sim/buttons/play-button-on.svg");
         background-color: rgba(0, 0, 0, 0);
         border: none;
         height: 48px;
         width: 48px;
 
         &:hover {
-          background-image: url("../assets/new-sim/buttons/play-button-hover.svg");
+          background-image: url("../../assets/new-sim/buttons/play-button-hover.svg");
         }
 
         &.playing {
-          background-image: url("../assets/new-sim/buttons/pause-button-on.svg");
+          background-image: url("../../assets/new-sim/buttons/pause-button-on.svg");
 
           &:hover {
-            background-image: url("../assets/new-sim/buttons/pause-button-hover.svg");
+            background-image: url("../../assets/new-sim/buttons/pause-button-hover.svg");
           }
         }
       }
@@ -173,23 +173,5 @@
 
   .person {
     position: absolute;
-  }
-
-  .simulation-button {
-    background-color: white;
-    border: 3px solid black;
-    border-radius: 6px;
-    bottom: 11px;
-    height: 54px;
-    position: absolute;
-    width: 55px;
-
-    &.reset {
-      right: 76px;
-    }
-
-    &.key {
-      right: 11px;
-    }
   }
 }
