+++ src/components/simulation-app.scss
@@ -72,30 +72,22 @@
       }
     }
   }
-}
-
-.simulation-footer {
-  display: flex;
-  gap: 10px;
-  margin: 0 5px 9px 9px;
-
-  .outcome-area {
-    align-items: center;
-    background-color: vars.$sim-transparent-background;
-    border-radius: 6px;
-    display: flex;
-    font-size: 18px;
-    height: 54px;
-    justify-content: center;
-    margin-right: 2px;
-    width: 740px;
-  }
 
   .simulation-button {
     background-color: white;
     border: 3px solid black;
     border-radius: 6px;
+    bottom: 11px;
     height: 54px;
+    position: absolute;
     width: 55px;
+
+    &.reset {
+      right: 76px;
+    }
+
+    &.key {
+      right: 11px;
+    }
   }
 }
